requirements.txt
opencv-python
numpy
Pillow
scikit-learn
matplotlib
